moved simple callbacks to telemetry_handler
